Use chain version of query proposal status
