+++ crates/cosmos/cosmos-relayer/src/contexts/chain.rs
@@ -1,5 +1,6 @@
 use alloc::sync::Arc;
 use core::ops::Deref;
+use hermes_test_components::chain::traits::proposal::query_status::CanQueryProposalStatus;
 use hermes_wasm_test_components::components::WasmChainComponents;
 use hermes_wasm_test_components::traits::chain::messages::store_code::StoreCodeMessageBuilderComponent;
 use hermes_wasm_test_components::traits::chain::upload_client_code::{
@@ -384,6 +385,7 @@ pub trait CanUseCosmosChain:
     + CanQueryTxResponse
     + CanAssertEventualAmount
     + CanUploadWasmClientCode
+    + CanQueryProposalStatus
 {
 }
