+++ crates/cosmos/cosmos-test-components/src/chain/impls/proposal/mod.rs
@@ -1,1 +1,2 @@
+pub mod poll_status;
 pub mod query_status;
