+++ crates/test/test-components/src/chain/traits/proposal/mod.rs
@@ -1,2 +1,3 @@
+pub mod poll_status;
 pub mod query_status;
 pub mod types;
