+++ crates/test/test-components/src/chain_driver/traits/proposal/mod.rs
@@ -1,4 +1,2 @@
 pub mod deposit;
-pub mod poll_status;
-pub mod query_status;
 pub mod vote;
